refactor(ViewModels): Switch location of credentials
